Get smoke-test assertions working in DEBUG mode.
Normally Karma runs all of its tests in an iframe. This allows it to
sandbox the test environment from the actuall test runner inside the
main Karma frame. However in debug mode, so that you can see the app
and also run devtools it runs your tests in the main frame. However,
you are subject to the main frames CSS as well as any global state
that might be hanging around between test runs.

In this instance, the TodoMVC page object was looking for the first
`h1` it found as the application's main heading. That was fine when it
was running in a sandbox, but when running in the main frame it was
finding Karma's navbar instead.

This adds a special class to the main React app so that the
application interactor can be scoped to that selector. That way, it
will work no matter which context it's in (unless there are multiple
apps running at the same time, and in that case it will select the first)